requirements.txt
twitchio>=2.8.0,<3.0
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.4
aiosqlite>=0.20.0
python-dotenv>=1.0.1
httpx>=0.27.0
